requirements.txt
django==5.0
djangorestframework==3.14.0
drf-spectacular==0.27.0
djoser==2.2.2
django-cors-headers==4.3.1
python-dotenv==1.0.0
redis==5.0.1
hiredis==2.2.3